+++ bothub_client/bot.py
@@ -13,7 +13,7 @@ class BaseBot(object):
         raise NotImplementedError()
 
     def send_message(self, message, user_id=None, channel=None):
-        self.channel_client.send_message(user_id, message, channel)
+        self.channel_client.send_message(user_id, message, channel, event=self.event)
 
     def set_project_data(self, data):
         self.storage_client.set_project_data(data)